+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/experience/enchanting/ISEEnchantingTableBlockEntity.java
@@ -98,13 +98,14 @@ public class ISEEnchantingTableBlockEntity extends BaseContainerBlockEntity impl
     }
 
     public boolean teachEnchantment(Enchantment enchantment, int lvl) {
-        if (this.knowsEnchantment(enchantment, lvl)) {
-            if (enchantment.getMaxLevel() == 1
+        if (this.knowsEnchantment(enchantment, lvl) && this.knownEnchantments.get(enchantment) >= lvl) {
+            return false;
+            /*if (enchantment.getMaxLevel() == 1
                     || this.knownEnchantments.get(enchantment) > lvl
                     || this.knownEnchantments.get(enchantment) == enchantment.getMaxLevel())
                 return false;
             this.knownEnchantments.put(enchantment, lvl + 1);
-            return true;
+            return true;*/
         }
         this.knownEnchantments.put(enchantment, lvl);
         this.setChanged();